python/requirements.txt
Flask == 3.0.2
PyPDF2 == 3.0.1
pdfminer.six == 20221105
pdfplumber == 0.10.3
pdf2image == 1.17.0
pillow == 10.2.0
pytesseract == 0.3.10
spacy == 3.7.2
en-core-web-sm==3.7.1
#or python -m spacy download en 